refactoring: moving some common functionality to JGitCommon
